Make task nodes outdated and rename maneuvering to task
