+++ comp_tasks/config/params.yaml
@@ -6,7 +6,7 @@ task:
     start_lon: 10.25448 #149.1654373
     finish_lat: 63.26445
     finish_lon: 10.25413
-maneuvering:
+task:
   ros__parameters:
     distance_to_move: 30.0 # meters
     angle_from_target: 20.0 # degrees
